improve txn orchestrator metrics

+++ crates/sui-core/src/transaction_driver/mod.rs
@@ -168,7 +168,7 @@ where
 
         for name in validators {
             let display_name = auth_agg.get_display_name(&name);
-            let delay_ms = rand::thread_rng().gen_range(0..max_jitter.as_millis());
+            let delay_ms = rand::thread_rng().gen_range(0..max_jitter.as_millis()) as u64;
             let self_clone = self.clone();
 
             let task = async move {
